New support for tracking creation times
Major refactor of dettraceSystemCalls.cpp
Lose dependency on std::filesystem
Use int return value to avoid signed comparison errors.
Update test submodule to passing status.
Minor documentation additions
Correctly mount /proc/ inside chroot.

+++ include/ptracer.hpp
@@ -243,11 +243,11 @@ public:
   void setReturnRegister(uint64_t retVal);
 
   /**
-   * Get results of system calls.
-   * During post system call event.
+   * Get results of system calls, we cast register value into int to avoid
+   * issues with sign.
    * @return Return value
    */
-  uint64_t getReturnValue();
+  int getReturnValue();
 
   /**
    * Get system call number.
